Fix: 좋아요 기능 수정
