--- src/domain/likes/entities/like.entity.ts
@@ -1,23 +1,0 @@
-import {
-  Column,
-  Entity,
-  JoinColumn,
-  ManyToOne,
-  PrimaryGeneratedColumn,
-} from 'typeorm';
-import { Post } from '@/domain/posts/entities/post.entity';
-
-@Entity({ name: 'likes' })
-export class Like {
-  @PrimaryGeneratedColumn('increment')
-  id: number;
-
-  @Column({ name: 'user_id', nullable: false })
-  userId: string;
-
-  @ManyToOne('Post', 'likes', {
-    onDelete: 'CASCADE',
-  })
-  @JoinColumn({ name: 'post_id', referencedColumnName: 'id' })
-  post: Post;
-}
