+++ src/domain/posts/posts.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { PostsService } from './posts.service';
-import { PostsController } from './posts.controller';
+import { LikesController, PostsController } from './posts.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Post } from './entities/post.entity';
 import { Repository } from 'typeorm';
@@ -10,7 +10,7 @@ import { SearchModule } from '@/domain/search/search.module';
 @Module({
   imports: [TypeOrmModule.forFeature([Post]), RedisCacheModule, SearchModule],
   exports: [PostsService, Repository<Post>],
-  controllers: [PostsController],
+  controllers: [PostsController, LikesController],
   providers: [PostsService, Repository<Post>],
 })
 export class PostsModule {}
